refactored model classes into models python files
