+++ mutton/__init__.py
@@ -2,5 +2,4 @@
 # -*- coding: utf-8 -*-
 
 from .base_handler import *
-from .base_request import *
-from .base_response import *
+from .models import *
