+++ mutton/apig/__init__.py
@@ -2,5 +2,4 @@
 # -*- coding: utf-8 -*-
 
 from .api_gateway_handler import *
-from .api_gateway_request import *
-from .api_gateway_response import *
+from .models import *
