+++ mutton/sqs/__init__.py
@@ -1,7 +1,5 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
+from .models import *
 from .sqs_event_handler import *
-from .sqs_event_request import *
-from .sqs_event_response import *
-from .sqs_message import *
